fixes for travis
fixes for travis
