+++ csci_2019_project/users/tests/test_models.py
@@ -1,8 +1,47 @@
 import pytest
 from django.conf import settings
+from typing import Any, Sequence
+from django.contrib.auth import get_user_model
+from factory import DjangoModelFactory, Faker, post_generation
 
 pytestmark = pytest.mark.django_db
 
+
+class UserFactory(DjangoModelFactory):
+
+    username = Faker("user_name")
+    email = Faker("email")
+    name = Faker("name")
+
+    @post_generation
+    def password(self, create: bool, extracted: Sequence[Any], **kwargs):
+        password = Faker(
+            "password",
+            length=42,
+            special_chars=True,
+            digits=True,
+            upper_case=True,
+            lower_case=True,
+        ).generate(extra_kwargs={})
+        self.set_password(password)
+
+    class Meta:
+        model = get_user_model()
+        django_get_or_create = ["username"]
+
+@pytest.fixture(autouse=True)
+def media_storage(settings, tmpdir):
+    settings.MEDIA_ROOT = tmpdir.strpath
+
+
+@pytest.fixture
+def user() -> settings.AUTH_USER_MODEL:
+    return UserFactory()
+
+
+@pytest.fixture
+def request_factory() -> RequestFactory:
+    return RequestFactory()
 
 def test_user_get_absolute_url(user: settings.AUTH_USER_MODEL):
     assert user.get_absolute_url() == f"/users/{user.username}/"
